feat(compile): Feature guard around nvim
If user has a shared configuration for nvim and vim this feature guard
will check if the user is on nvim if they are not then it will just
`finish` sourcing the file.

Users might want to have the same configuration for both neovim and vim.
If for some reason they are on vim instead of nvim their configuration
will gracefully degrate intead of throwing errors.

As the compiled file vimscript file is in the users runtime it should
degrade gracefully.