ref(clean): use find_missing_plugins to implement
Previously, clean was implemented using a lot of repeated logic which
could be discerned from other parts of the program. This commit
reimplements it in terms of these other portions, such as
`find_missing_plugins` and `plugin.install_path`.

This transforms it into a top-down approach rather than a bottom-up
approach: we assume that by scanning every specified plugin, we can
determine each folder which belongs to a configuration. Then, each
folder not belonging to a configuration must be 'dirty' and needs
cleaning.

Before, we analyzed the folders and attempted to then map them to the
plugins. This took a lot more work and made it so that logic for
aliasing had to be duplicated across files. This will resolve that and
any future changes automatically.