fix(plugin_utils): find_missing_plugins by dir type
In order to find out if a specific directory has been
changed and needs cleaning, there needed to be a function which detects
a directory's state from nothing more than the path.

After all, it is completely possible that the user has gone into the
directory and made changes. As a matter of fact, in testing this
function it managed to catch a few mismatches that I myself had
neglected to specify in my plugin spec.

PR #143 did not catch these mismatches so I'm inclined to believe
that this is a better implementation of what we're trying to do.