fix(clean): detect uninstalled plugins
The previous implementation only detected when there was a
dirty/missing plugin. It was missing the logic for knowing whether or
not the directories actually corresponded to a specified plugin.

This approach theoretically will be more responsive to changes in the
rest of the plugin, since it uses more functions from other places. For
example, any new method for detecting changed/missing plugins will
automatically apply to the clean method now.