Use aliased name as plugin name if present (#142)
* Use aliased name as plugin name if present

This is in reference to #140. Before, the `as` name was not used preferentially to the repo name provided in the specification. This will make the `as` alias be used instead, if present.

* Remove aliasing logic from short_name

short_name used to have plugin.as as part of its logic, but now that plugin.as is used earlier in the function it isn't needed here again.