Check dir type for changes (#144)
* fix(plugin_utils): find_missing_plugins by dir type

In order to find out if a specific directory has been
changed and needs cleaning, there needed to be a function which detects
a directory's state from nothing more than the path.

After all, it is completely possible that the user has gone into the
directory and made changes. As a matter of fact, in testing this
function it managed to catch a few mismatches that I myself had
neglected to specify in my plugin spec.

PR #143 did not catch these mismatches so I'm inclined to believe
that this is a better implementation of what we're trying to do.

* fix(clean): implement with find_missing_plugins

Previously, clean was implemented using a lot of repeated logic which
could be discerned from other parts of the program. This commit
reimplements it in terms of these other portions, such as
`find_missing_plugins` and `plugin.install_path`.

This transforms it into a top-down approach rather than a bottom-up
approach: we assume that by scanning every specified plugin, we can
determine each folder which belongs to a configuration. Then, each
folder not belonging to a configuration must be 'dirty' and needs
cleaning.

Before, we analyzed the folders and attempted to then map them to the
plugins. This took a lot more work and made it so that logic for
aliasing had to be duplicated across files. This will resolve that and
any future changes automatically.

* ref: use plugin type constants

The previous commit defined constants which contain the plugin type
names. Rather than just using strings I figured it would be better to
use them consistently.

* ref(clean): use constants consistently

* fix(clean): detect uninstalled plugins

The previous implementation only detected when there was a
dirty/missing plugin. It was missing the logic for knowing whether or
not the directories actually corresponded to a specified plugin.

This approach theoretically will be more responsive to changes in the
rest of the plugin, since it uses more functions from other places. For
example, any new method for detecting changed/missing plugins will
automatically apply to the clean method now.

* ref: rename XXX_plugin to XXX_plugin_type

* style: replace literal \t with '  '

* ref(clean): use `table.insert` over `#foo+1`

See the relevant thread about this.

SEE: https://github.com/wbthomason/packer.nvim/pull/144/files#r551995945