add configuration option 'wants' for plugins (#264)
The following use cases needs to be solved. The plugin A is depending on
plugin B. A and B are both plugins maintained by Packer, so B is not
"just" a 'requires' plugin of A. When ever A gets loaded for what ever
trigger (might be lazy loaded), first plugin B must be loaded to make A
working.
This is pretty stupid simple first version implementing this feature.
Therefore it adds a new plugin configuration property called 'wants'.
When a plugin is about to get loaded, it will be checked if this plugin
wants any other plugins loaded. If so, first all these plugins get
loaded before the loading process of the actual plugin continues. If a
wanted plugin is already loaded, nothing happens as always.