Fix highlight being disabled due to redrawtimeout
According to docs '\@<=' is really inefficient and recommends using `\zs`
instead of it . I've replace the `\@<=` with `\zs` in syntax rules.

For packerStatusSuccess and packerStatusFail I've limited their look
behind to length of look-behind string . Couldn't get them to work with `\zs` . I'm assuming
it's because both `^ working_sym` is already matches packerSuccess.