feat(packer): remove VimL packer_compiled if `output_lua`
This commit adds functionality which removes a user's old
'packer_compiled' VimL file if they compile it with Lua. It does so by
looking for a file of the same name but different extension (e.g.
'packer_compiled.lua' & 'packer_compiled.vim') and removes the VimL
variant. This only occurs if the files are in the same directory, so if
a user alters their `compile_path` when changing to Lua it will not be
detected (e.g. 'plugin/packer_compiled.vim' -> 'plugin/packer.lua' won't
be caught).