Fix #469: Avoid re-running module loaders for lazy-loaded modules
This was a fun one! #469 was caused because the `module` lazy-loader had
the following control flow:
1. Call `packer.load` for the plugins that needed to be loaded
2. In `packer.load`, run any `config` key functions, **which may require
   the lazy-loaded module**.
3. Return `nil`, to pass on the loading process to a subsequent module
   loader.

The problem is that, in step 2, the lazy-loader may get recursively
invoked but (correctly) skip to step 3, actually loading the module.
Then, by doing step 3 **again** in the original invocation, we basically
force the package.path/other loader to run again, re-loading the module.

This commit fixes the issue by instead returning the **already loaded**
module from `package.loaded`, if indeed this value exists.

Closes #469