Document the structure of packer.update
There is a ton of code duplication between major Packer operations
(install, update, sync). The first step to refactoring is learning
what makes it tick. I made some notes that outline the flow of the
`packer.update` function, which is mostly mirrored in the rest of
the operations as well. I also marked which parts are the same and
which are different between `packer.update` and `packer.sync`.