README.md: Don't scare reader with crufty notices
Notices about year-old bugs or other historical details are not a great
intro text. Best to hide that at the end (or remove it entirely).

Similarly, let's minimize the emphasis on Neovim 0.5 min version; it's
not new, and no one is running anything older at this point.

Similarly, remove old "current work in progress" status message which is
impossible to keep fresh.

Probably more cleanup is warranted, but this is what struck my new-user
eyes.