feat: create singleton events/commands/keymaps for load handlers
Implementing the idea @lewis6991 proposed for handling multiple loads on
the same command/keymap/event added by separate calls to `add`. Changes
the loaders to keep a table of plugins associated with a given load
trigger and only create new events/keymaps/commands for new triggers,
otherwise simply appending to the set of plugins to be loaded on the
given trigger.

One note: we may need to make loaders delete plugins from these lists
after load, to handle cases like:
- add(event, plugins_1)
- event happens, plugins_1 loaded
- add(event, plugins_2)
- event happens, without deleting plugins_1 gets loaded again

I don't believe that this is a potential bug introduced by this
change---it could've happened with the more naive handlers as well