Fix issues with the packer.init config example
There were two issues I noticed with this. One was invalid function calls (`stdpath` instead of `vim.fn.stdpath`) and using undefined modules (`util.join_paths`).

The other was having the table represented in absence of context. Am I meant to pass it to packer.init()? Return it like the example from packer.startup()?

I've fixed the functions and added the needed require statements. I've also added how I believe it is meant to be passed to the packer.init() making it a drop in example.