Clarify installation instructions
The old instructions said that the Git repository needs to be cloned
somewhere on your `packpath`. However, it actually needs to be installed
in the `pack/FOOBAR/start/` subdirectory of a directory in `packpath`,
replacing `FOOBAR` with any name. This fix will prevent confusion for
users who are still learning `packpath` and who may have their
configuration in `~/.config/nvim/` rather than
`~/.local/share/nvim/site/` .

This commit also fixes the documentation in `doc/packer.txt` to match
`README.md` . In particular, the mention of the `opt` directory was
fixed, it now mentions the `start` directory.